% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{gpg_encrypt}
\alias{gpg_decrypt}
\alias{gpg_encrypt}
\title{Encryption}
\usage{
gpg_encrypt(data, receiver, signer = NULL)

gpg_decrypt(data, verify = TRUE)
}
\arguments{
\item{data}{path to file or raw vector with data to encrypt / decrypt}

\item{receiver}{one or more key ids or fingerprints for recepient(s)}

\item{signer}{(optional) key ids or fingerprints for the sender to sign the message}

\item{verify}{automatically checks that all signatures can be verified and raises an
error otherwise}
}
\description{
Encrypt or decrypt a message using the public key from the \code{receiver}.
Optionally the message can be signed using the private key of the sender.
}

