% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen.R
\name{gpg_keygen}
\alias{gpg_keygen}
\title{GPG keygen}
\usage{
gpg_keygen(name, email, passphrase = NULL, ...)
}
\arguments{
\item{name}{value for the \code{Name-Real} field}

\item{email}{value for the \code{Name-Email} field}

\item{passphrase}{(optional) protect with a passphrase}

\item{...}{other fields, see \href{https://www.gnupg.org/documentation/manuals/gnupg/Unattended-GPG-key-generation.html}{GPG manual}

[GPG manual]: R:GPG manual}
}
\description{
Generates a new private-public keypair. Note that some options are
only supported if you have GPG2. The \href{https://www.gnupg.org/documentation/manuals/gnupg/Unattended-GPG-key-generation.html}{GPG manual}
explains which fields are suppor

[GPG manual]: R:GPG manual
}

