% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{gpg_options}
\alias{gpg_options}
\title{GPG options}
\usage{
gpg_options(...)
}
\description{
Get and set options GnuPG configuration options.
}
\details{
This function reads and edits options in the GnuPG configuration file, which
is stored by default in \code{~/.gnupg/gpg.conf}. Note that changing options
can affect other software using this installation of GnuPG.
}
\examples{
# Read options:
gpg_options()
gpg_options("keyserver")

# Write options
gpg_options("debug-level" = "9")
gpg_options("debug-level" = "none")
}
\seealso{
Other gpg: \code{\link{gpg}}, \code{\link{gpg utilities}},
  \code{\link{gpg_import}}, \code{\link{gpg_list}},
  \code{\link{gpg_sign}}, \code{\link{gpg_verify}}
}

