% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart.R
\name{gpg_restart}
\alias{gpg_restart}
\title{Start GPG engine}
\usage{
gpg_restart(path = NULL, home = NULL, wininst = NULL, debug = "none")
}
\arguments{
\item{path}{location of \code{gpg} or \code{gpg2} or \code{gpgconf}}

\item{home}{path to your GPG configuration directory}

\item{wininst}{path to \code{gpgme-w32spawn.exe} executable on Windows}

\item{debug}{debugging level, integer between 1 and 9}
}
\description{
Finds the \code{gpg} program and loads a context.
}

