% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{pinentry}
\alias{pinentry}
\title{Password Entry}
\usage{
pinentry(prompt = "Enter your GPG passphrase:")
}
\arguments{
\item{prompt}{the string printed when prompting the user for input.}
}
\description{
Function to prompt the user for a password to read a protected private key.
}
\details{
If available, this function calls the GnuPG \code{pinentry} program. However this
only works in a terminal. Therefore the IDE can provide a custom password entry
widget by setting the \code{askpass} option. If no such option is specified
we default to \code{\link{readline}}.
}

