% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verify.R
\name{gpg utilities}
\alias{gpg}
\alias{gpg utilities}
\alias{gpg_verify}
\title{PGP tools}
\usage{
gpg_verify(sigfile, datafile)
}
\arguments{
\item{sigfile}{path to the gpg file containing the \code{PGP SIGNATURE} block.}

\item{datafile}{path to the file containing the message to be verified.}
}
\description{
Utilities to create and verify PGP signatures.
}
\examples{
# This requires you have the Debian master key in your keyring
# See https://lists.debian.org/debian-devel-announce/2014/11/msg00017.html
# gpg --keyserver pgp.mit.edu --recv 0x7638d0442b90d010
msg <- tempfile()
sig <- tempfile()
download.file("http://http.us.debian.org/debian/dists/jessie/Release", msg)
download.file("http://http.us.debian.org/debian/dists/jessie/Release.gpg", sig)
gpg_verify(sig, msg)
}

