% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R, R/options.R, R/restart.R, R/signing.R
\name{gpg_import}
\alias{gpg}
\alias{gpg utilities}
\alias{gpg_delete}
\alias{gpg_import}
\alias{gpg_info}
\alias{gpg_list_keys}
\alias{gpg_list_secret_keys}
\alias{gpg_recv}
\alias{gpg_restart}
\alias{gpg_sign}
\alias{gpg_verify}
\alias{gpg_version}
\title{Start GPG engine}
\usage{
gpg_import(pubkey)

gpg_recv(id, keyserver = "https://keyserver.ubuntu.com")

gpg_list_keys()

gpg_list_secret_keys()

gpg_delete(id, allow_secret = TRUE)

gpg_info()

gpg_version(silent = FALSE)

gpg_restart(path = NULL, home = NULL, debug = "none", silent = FALSE)

gpg_verify(sigfile, datafile)

gpg_sign(datafile, name = "", password = readline("ENTER PASSWORD: "))
}
\arguments{
\item{pubkey}{path to a file with a trusted public key}

\item{id}{unique ID of the pubkey (starts with \code{0x})}

\item{keyserver}{address of http keyserver}

\item{allow_secret}{set to TRUE if are sure you want to a private keys}

\item{silent}{suppress output of \code{gpg --version}}

\item{path}{location of \code{gpg} or \code{gpg2} or \code{gpgconf} or (on windows) \code{gpgme-w32spawn.exe}}

\item{home}{path to your GPG configuration directory}

\item{debug}{debugging level, integer between 1 and 9}

\item{sigfile}{path to the gpg file containing the \code{PGP SIGNATURE} block.}

\item{datafile}{path to the file containing the message to be verified.}

\item{name}{find a key that matches a particular name}

\item{password}{a string or expression callback to read a passphrase when needed}
}
\description{
Finds the \code{gpg} program and loads a context.

Utilities to create and verify PGP signatures.
}
\examples{
gpg_info()
gpg_version()
# This requires you have the Debian master key in your keyring
# See https://lists.debian.org/debian-devel-announce/2014/11/msg00017.html
# gpg --keyserver pgp.mit.edu --recv 0x7638d0442b90d010
msg <- tempfile()
sig <- tempfile()
download.file("http://http.us.debian.org/debian/dists/jessie/Release", msg)
download.file("http://http.us.debian.org/debian/dists/jessie/Release.gpg", sig)
gpg_verify(sig, msg)
}
\seealso{
Other gpg: \code{\link{gpg_options}},
  \code{\link{gpg_search}}
}

