% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R, R/signing.R
\name{gpg_import}
\alias{gpg}
\alias{gpg utilities}
\alias{gpg_import}
\alias{gpg_list}
\alias{gpg_sign}
\alias{gpg_verify}
\title{PGP tools}
\usage{
gpg_import(pubkey)

gpg_list(secret_only = FALSE)

gpg_verify(sigfile, datafile)

gpg_sign(datafile, name = "", password = readline("ENTER PASSWORD: "))
}
\arguments{
\item{pubkey}{path to a file with a trusted public key}

\item{secret_only}{only list keys for which we have the secret (private key)}

\item{sigfile}{path to the gpg file containing the \code{PGP SIGNATURE} block.}

\item{datafile}{path to the file containing the message to be verified.}

\item{name}{find a key that matches a particular name}

\item{password}{a string or expression callback to read a passphrase when needed}
}
\description{
Utilities to create and verify PGP signatures.
}
\examples{
# This requires you have the Debian master key in your keyring
# See https://lists.debian.org/debian-devel-announce/2014/11/msg00017.html
# gpg --keyserver pgp.mit.edu --recv 0x7638d0442b90d010
msg <- tempfile()
sig <- tempfile()
download.file("http://http.us.debian.org/debian/dists/jessie/Release", msg)
download.file("http://http.us.debian.org/debian/dists/jessie/Release.gpg", sig)
gpg_verify(sig, msg)
}
\seealso{
Other gpg: \code{\link{gpg_options}},
  \code{\link{gpg_search}}
}

