% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{gpg_encrypt}
\alias{gpg_decrypt}
\alias{gpg_encrypt}
\title{Encryption}
\usage{
gpg_encrypt(file, id)

gpg_decrypt(file)
}
\arguments{
\item{file}{path to file or raw vector with data to encrypt / decrypt}

\item{id}{key id or fingerprint}
}
\description{
Encrypt or decrypt a message.
}

