% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen.R
\name{gpg_keygen}
\alias{gpg_keygen}
\title{GPG key generation}
\usage{
gpg_keygen(name, email, passphrase = NULL)
}
\arguments{
\item{name}{value for the \code{Name-Real} field}

\item{email}{value for the \code{Name-Email} field}

\item{passphrase}{(optional) protect with a passphrase}
}
\description{
Generates a new standard private-public keypair. This function is mostly
for testing purposes. Use the \code{gpg --gen-key} command line utility to generate an
official GPG key with custom fields and options.
}
\references{
GPG manual section on
\href{https://www.gnupg.org/documentation/manuals/gnupg/Unattended-GPG-key-generation.html}{Unattended key generation}.
}
\seealso{
Other gpg: \code{\link{gpg_encrypt}},
  \code{\link{gpg_keys}}, \code{\link{gpg_sign}}
}

