% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygen.R
\name{gpg_keygen}
\alias{gpg_keygen}
\title{GPG key generation}
\usage{
gpg_keygen(name, email, passphrase = NULL, key_type = "RSA", ...)
}
\arguments{
\item{name}{value for the \code{Name-Real} field}

\item{email}{value for the \code{Name-Email} field}

\item{passphrase}{(optional) protect with a passphrase}

\item{key_type}{required field, defaults to RSA}

\item{...}{other fields, see \href{https://www.gnupg.org/documentation/manuals/gnupg/Unattended-GPG-key-generation.html}{GPG manual}

[GPG manual]: R:GPG manual}
}
\description{
Use \code{gpg_keygen} to generate a new private-public keypair. Supported options
and parameters depend on the version of GPG. See the GPG manual section on
\href{https://www.gnupg.org/documentation/manuals/gnupg/Unattended-GPG-key-generation.html}{Unattended key generation}.

[Unattended key generation]: R:Unattended key generation
}

