% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{gpg_import}
\alias{gpg_delete}
\alias{gpg_import}
\alias{gpg_list_keys}
\alias{gpg_list_secret_keys}
\alias{gpg_recv}
\title{GPG keyring management}
\usage{
gpg_import(file)

gpg_recv(id, keyserver = "https://keyserver.ubuntu.com")

gpg_delete(id, allow_secret = TRUE)

gpg_list_keys()

gpg_list_secret_keys()
}
\arguments{
\item{file}{path to the key file or raw vector with key data}

\item{id}{unique ID of the pubkey (starts with \code{0x})}

\item{keyserver}{address of http keyserver}

\item{allow_secret}{set to TRUE if are sure you want to a private keys}
}
\description{
Signing or encrypting with GPG require that the keys are stored in your
personal keyring. Use \link{gpg_version} to see which keyring (home dir)
you are using. Also see \link{gpg_keygen} for generating a new key.
}

