% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpg.R
\name{gpg_search}
\alias{gpg_download}
\alias{gpg_search}
\title{GPG keyserver}
\usage{
gpg_search(name = "")

gpg_download(id = "")
}
\description{
Search, download or publish keys on the GPG keyserver network via the HKP
(OpenPGP HTTP Keyserver Protocol). Requires a valid keyserver is configured
in your gpg.conf file. Might not work behind a firewall.
}

