% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{gpg_search}
\alias{gpg_search}
\title{GPG keyserver}
\usage{
gpg_search(name = "")
}
\arguments{
\item{name}{filter keys by name}
}
\description{
Search, download or publish keys on the GPG keyserver network via the HKP
(OpenPGP HTTP Keyserver Protocol). Requires a valid keyserver is configured
in your gpg.conf file. Might not work behind a firewall.
}
\seealso{
Other gpg: \code{\link{gpg utilities}}
}

