% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gpg_options}
\alias{gpg_options}
\title{GPG options}
\usage{
gpg_options()
}
\description{
Get and set options GnuPG configuration options.
}
\details{
This function reads and edits options in the GnuPG configuration file, which
is stored by default in \code{~/.gnupg/gpg.conf}. Note that changing options
can affect other software using this installation of GnuPG.
}
\examples{
# Read options:
gpg_options()
}
\seealso{
Other gpg: \code{\link{gpg_import}},
  \code{\link{gpg_search}}
}

