% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signing.R
\name{gpg_verify}
\alias{gpg}
\alias{gpg_sign}
\alias{gpg_verify}
\title{PGP Signatures}
\usage{
gpg_verify(file, signature, error = TRUE)

gpg_sign(file, id = NULL)
}
\arguments{
\item{file}{file-path or raw vector with data to sign or verify}

\item{signature}{path or raw vector for the gpg signature file (contains the \code{PGP SIGNATURE} block)}

\item{error}{raise an error if verification fails because you do not have the
signer public key in your keyring.}

\item{id}{(optional) vector with key ID's to use for signing. If \code{NULL}, GPG defaults
to first private key in the keyring, or what has been configured in global options.}
}
\description{
Utilities to create and verify PGP signatures.
}
\examples{
# This requires you have the Debian master key in your keyring
# See https://lists.debian.org/debian-devel-announce/2014/11/msg00017.html
msg <- tempfile()
sig <- tempfile()
download.file("http://http.us.debian.org/debian/dists/jessie/Release", msg)
download.file("http://http.us.debian.org/debian/dists/jessie/Release.gpg", sig)
gpg_verify(msg, sig, error = FALSE)
}
\seealso{
Other gpg: \code{\link{gpg_search}}
}

